volta list: extract `list::Toolchain` to own module.
